src/functions/risk_prediction/requirements.txt
functions-framework==3.2.0
google-cloud-storage==2.19.0
python-dotenv==1.0.1
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.3.2
